requirements.txt
streamlit>=1.45.1
groq>=0.27.0
mistralai>=1.8.2
pandas>=2.3.0
pillow>=11.2.1
python-dotenv>=1.1.0
requests>=2.32.4
pydantic>=2.11.0
openai
psycopg2-binary
spacy
PyMuPDF
sentence-transformers
python-docx
